+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - [jail] feature: `jailed players` now have their `display names` modified.
 - [command_toolbox.nickname] feature: improve the compatibility with other mods.
 - [world] refactor: remove the `blacklist` option to reduce confusion, as its functionality is covered by the `--confirm`.
+- [placeholder] feature: now the `%fuji:player_{prefix|suffix}%` placeholder returns a dummy `empty string` if the specified player has no `prefix` or `suffix`. (Makes it easier to use)